refactor(ui): move generateParams to a util
Move the util for preparing payload parameters to a util.

Fixes: canonical-web-and-design/app-tribe#556.
