+++ ui/src/app/store/machine/slice.ts
@@ -57,25 +57,7 @@ import {
   genericInitialState,
 } from "app/store/utils/slice";
 import type { StatusHandlers } from "app/store/utils/slice";
-import { kebabToCamelCase } from "app/utils";
-
-const generateParams = <P extends { [x: string]: unknown }>(
-  params: P,
-  mapping: { [x: string]: string } = {}
-) => {
-  const payload: { [x: string]: unknown } = {};
-  Object.entries(params).forEach(([key, value]) => {
-    if (key in mapping) {
-      // if the payload should use a different key then update it.
-      key = mapping[key];
-    }
-    // Don't include any undefined values.
-    if (typeof value !== "undefined") {
-      payload[key] = value;
-    }
-  });
-  return payload;
-};
+import { generatePayloadParams, kebabToCamelCase } from "app/utils";
 
 export const ACTIONS: Action[] = [
   {
@@ -570,7 +552,7 @@ const machineSlice = createSlice({
           method: "create_bcache",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             cacheMode: "cache_mode",
             cacheSetId: "cache_set",
             systemId: MachineMeta.PK,
@@ -595,7 +577,7 @@ const machineSlice = createSlice({
           method: "create_bond",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -612,7 +594,7 @@ const machineSlice = createSlice({
           method: "create_bridge",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -629,7 +611,7 @@ const machineSlice = createSlice({
           method: "create_cache_set",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockId: "block_id",
             partitionId: "partition_id",
             systemId: MachineMeta.PK,
@@ -650,7 +632,7 @@ const machineSlice = createSlice({
           method: "create_logical_volume",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             mountOptions: "mount_options",
             mountPoint: "mount_point",
             systemId: MachineMeta.PK,
@@ -709,7 +691,7 @@ const machineSlice = createSlice({
           method: "create_physical",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -726,7 +708,7 @@ const machineSlice = createSlice({
           method: "create_raid",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockDeviceIds: "block_devices",
             mountOptions: "mount_options",
             mountPoint: "mount_point",
@@ -751,7 +733,7 @@ const machineSlice = createSlice({
           method: "create_vlan",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -768,7 +750,7 @@ const machineSlice = createSlice({
           method: "create_vmfs_datastore",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockDeviceIds: "block_devices",
             partitionIds: "partitions",
             systemId: MachineMeta.PK,
@@ -789,7 +771,7 @@ const machineSlice = createSlice({
           method: "create_volume_group",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockDeviceIds: "block_devices",
             partitionIds: "partitions",
             systemId: MachineMeta.PK,
@@ -882,7 +864,7 @@ const machineSlice = createSlice({
           method: "delete_filesystem",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockDeviceId: "blockdevice_id",
             filesystemId: "filesystem_id",
             partitionId: "partition_id",
@@ -1128,7 +1110,7 @@ const machineSlice = createSlice({
           method: "link_subnet",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -1586,7 +1568,7 @@ const machineSlice = createSlice({
           method: "update_disk",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockId: "block_id",
             mountOptions: "mount_options",
             mountPoint: "mount_point",
@@ -1608,7 +1590,7 @@ const machineSlice = createSlice({
           method: "update_filesystem",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockId: "block_id",
             mountOptions: "mount_options",
             mountPoint: "mount_point",
@@ -1635,7 +1617,7 @@ const machineSlice = createSlice({
           method: "update_interface",
         },
         payload: {
-          params: generateParams(params),
+          params: generatePayloadParams(params),
         },
       }),
       reducer: () => {
@@ -1652,7 +1634,7 @@ const machineSlice = createSlice({
           method: "update_vmfs_datastore",
         },
         payload: {
-          params: generateParams(params, {
+          params: generatePayloadParams(params, {
             blockDeviceIds: "add_block_devices",
             partitionIds: "add_partitions",
             systemId: MachineMeta.PK,
