+++ ui/src/app/utils/index.ts
@@ -9,15 +9,16 @@ export { formatErrors } from "./formatErrors";
 export { formatMacAddress } from "./formatMacAddress";
 export { formatSpeedUnits } from "./formatSpeedUnits";
 export { generateCheckboxHandlers } from "./generateCheckboxHandlers";
+export { generatePayloadParams } from "./generatePayloadParams";
 export { getCookie } from "./getCookie";
 export { getNextName } from "./getNextName";
 export { getRanges } from "./getRanges";
 export { getStatusText } from "./getStatusText";
 export { groupAsMap } from "./groupAsMap";
 export { isComparable } from "./isComparable";
-export { objectHasKey } from "./objectHasKey";
 export { isVersionNewer } from "./isVersionNewer";
 export { kebabToCamelCase } from "./kebabToCamelCase";
+export { objectHasKey } from "./objectHasKey";
 export { preparePayload } from "./preparePayload";
 export { simpleSortByKey } from "./simpleSortByKey";
 export { someInArray } from "./someInArray";
